GR block dtype correctness: shim/YAML consistency gate + fix 4 blocks
Root-caused why real GNU Radio flowgraphs (SSB Weaver, BPSK modem) show red
port-dtype errors in GRC: the GR marker shims and their .block.yml port
declarations disagreed with the blocks' own VERIFIED GNU-Radio equivalents, and
NO test checked this (each block's verification drives the on-chip DUT via
run_block_dut_complex, bypassing the GR shim's stream dtype).

New gate (proven to FAIL first, INV-4): verification/tests/test_grc_block_port_dtypes.py
instantiates each real GR block, reads its io_signature, and asserts the shim AND
the .block.yml agree with each other and with the verified equivalent. Resolves a
${io_type} port dtype to the param default.

Fixes it caught + corrected (all proven by re-running the blocks' verification):
- complex_mixer shim: omitted in/out dtype -> defaulted to float; now complex->complex
  (matches verified multiply_cc). 25/25 test_complex_mixer green.
- iq_upconvert .block.yml: declared two float xi/xq rails; now ONE complex in ->
  float out (matches verified multiply_cc->complex_to_real + the real block's
  io_signature). 11/11 test_iq_upconvert green.
- rrc_pulse_shaper + upsampler shims: hardcoded complex64 while .block.yml io_type
  default is float; made shims io_type-aware (default float) + make: passes io_type.
  24/24 rrc+upsampler green.

Proven no regression: 91 gate+block tests, 24 rrc/upsampler, 17 BPSK modem
(test_bpsk_modem + grc_import + duplex_e2e) all green.

+++ gr-kyttar/python/kyttar/complex_mixer.py
@@ -12,6 +12,8 @@ Copyright 2026 Kyttar Computer Project.
 SPDX-License-Identifier: GPL-3.0-or-later
 """
 
+import numpy as np
+
 from .dsp_markers import _PassThrough
 
 
@@ -38,7 +40,13 @@ class complex_mixer(_PassThrough):
         offset: float = 0.0,
         phase: float = 0.0,
     ):
-        super().__init__(name="Kyttar Complex Mixer", n_in=1, n_out=1)
+        # COMPLEX -> COMPLEX, matching the verified GNU Radio equivalent
+        # multiply_cc(signal, sig_source_c) (see verification/tests/test_complex_mixer.py).
+        # The shim previously omitted the dtypes and defaulted to float, which
+        # disagreed with the block's own .block.yml (complex) and made a real GRC
+        # flowgraph reject the wiring — caught by test_grc_block_port_dtypes.
+        super().__init__(name="Kyttar Complex Mixer", n_in=1, n_out=1,
+                         in_dtype=np.complex64, out_dtype=np.complex64)
         self._device_id = device_id
         self._sample_rate = sample_rate
         self._frequency = frequency
